flowbit-analytics-assignment/services/vanna/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.3
requests==2.31.0
vanna[chromadb,openai]==0.5.5
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
openai==1.3.0
vanna==0.3.4
pydantic==2.5.0